refactor: move SSAReadTracker to _utils
